+++ wine/wine/library.h
@@ -180,14 +180,14 @@ static inline int wine_ldt_is_empty( const LDT_ENTRY *ent )
     static FORCEINLINE unsigned short wine_get_##seg(void) \
     { unsigned short res; __asm__ __volatile__("movw %%" #seg ",%w0" : "=r"(res)); return res; }
 #  define __DEFINE_SET_SEG(seg) \
-    static FORCEINLINE void wine_set_##seg(int val) \
+    static FORCEINLINE void wine_set_##seg(unsigned int val) \
     { __asm__("movw %w0,%%" #seg : : "r" (val)); }
 # elif !USE_SREG_WORKAROUND && defined(_MSC_VER)
 #  define __DEFINE_GET_SEG(seg) \
     static inline unsigned short wine_get_##seg(void) \
     { unsigned short res; __asm { mov res, seg } return res; }
 #  define __DEFINE_SET_SEG(seg) \
-    static inline void wine_set_##seg(unsigned short val) { __asm { mov seg, val } }
+    static inline void wine_set_##seg(unsigned int val) { __asm { mov seg, val } }
 # else  /* __GNUC__ || _MSC_VER */
 #  define __DEFINE_GET_SEG(seg) extern unsigned short wine_get_##seg(void);
 #  define __DEFINE_SET_SEG(seg) extern void wine_set_##seg(unsigned int);
